requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
env==0.1.0
idna==3.10
peewee==3.17.8
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
telebot==0.0.5
urllib3==2.3.0
